vm: initial asset lock time validation
This commit applies the functional options pattern to the `Engine`
struct. This allows for passing in the block height to the constructor
by using the `WithBlockHeight` option func.

Asset lock time is added to `validateStateTransition`. It
rejects assets as input if the asset has a `LockTime` other than zero
AND the block height of that the vm received from the constructor is
less than the assets `LockTime`.