multi: change lock time semantics
This commit changes the way the LockTime and RelativeLockTime fields are
interpreted. Because we cannot enforce that an asset sets these values
when it is being created for a certain script key that contains a time
lock (e.g. we might receive to an address for a script key that contains
a time lock, but the lock time and relative lock time values aren't part
of the address), we cannot expect an _input_ asset to have these values
set.
Instead, an asset attempting to spend from a time locked script key
needs to present the correct lock time values.