vm: make tests run with `WithTimeLockValidation`
This commit makes the tests run with `WithTimeLockValidation` enabled.
It also adds tests to check whether lock times are enforced correctly.

In `vm_test.go` if the `currentHeight` passed to `verifyTestCase` is
zero, that function will calculate `currentHeight` based off the lock
times of the inputs and use a block height that makes the inputs
validate.

`tapsend` tests that failed because of lock times now being
enforced are fixed.