tapfreighter: validate vPkt versions for parcel
In this commit, we add extra validation logic at the entrypoints to the
tapfreighter to assert that all vPackets in a received parcel are of a
matching version. This version determines the output anchor commitment
version, and V0 and V1 vPacket versions cannot be mixed. We also
propagate the active vPkt version to passive vPkts to keep vPkt versions
the same for all transfer outputs.