tapchannelmsg: fix empty msg formatting, add pending chans
This commit returns an empty string if there is no custom data in a
channel. That prevents the lncli output from showing:
  "custom_channel_data": "{\"assets\":null}",
And instead just shows:
  "custom_channel_data: "",

At the same time, we add the previously forgotten pending channels to
the calls of RPC methods that we want to format the response for.