rfq: round up inbound amount in policy check
To make sure the lnd forwarding algorithm doesn't reject an incoming
HTLC because due to the asset resolution it appears they aren't paying
enough fees, we always add a single asset unit.
This doesn't actually mean we're sending an additional unit every time,
it just means we might accept an HTLC even if it pays a couple of
milli-satoshis too little fees. This can be remediated by increasing the
asset resolution/precision.