rfq: add pure asset forwarding policy
If a payment is coming in through one asset channel and is supposed to
leave through another asset channel, we don't actually have to do any
conversion to BTC but instead to the outgoing asset directly (this would
be a single hop payment with asset channels on both sides).
To allow that forwarding mode, we add a new policy that contains both an
incoming and outgoing channel policy.