vm+tapsend: fix copy bug, fix lint issue, fix error
This commit mainly fixes a bug that when we copy an asset to create a
new output, we didn't empty out the split commitment when we spent an
output of a previous split in a full-value send output.
This lead to a validation error in the VM. We also add some more context
to the error that we were previously running into with this.
And just because we can, we fix a package shadowing issue.