tapchannel: import close output as address
If we're not the initiator of the shutdown process, we won't be called
back from lnd after the channel closing TX confirmed. So in order for us
to pick up on the transfer and download the proof from the courier, we
need to add our expected output as a TAP address, so the custodian will
detect the confirmed output on chain.

This is not optimal, as it will not work if there are multiple assets in
that close output. But since we'll need to add a whole new callback in
lnd that invokes a method in the aux closer after the channel confirmed,
we solve it this way instead for now. Because the whole funding logic
isn't ready to add multiple assets to a channel anyway.