tapfreighter: verify proof before import
In the freighter, we import the proof into the file based store only,
which doesn't do proof validation (instead of the multi archiver which
does verify first), due to the way we log and confirm parcels.
But that means we didn't verify proofs after creating them, which can
lead to issues.
By explicitly verifying we make sure we only import proper proofs.