proof: add helpers for deciaml display in metadata
In this commit, we add helpers to further validate metadata size,
and get and set the decimal display field. The default decimal display
value is 0, even if not encoded. We have to check the size of the
metadata if we add decimal display as a new field. We also don't want to
clobber JSON with an existing value with the same key, that may not be a
valid decimal display value.