tapchannelmsg: add max read size argument
This commit introduces a `max read size` argument to the
`ReadOpenChannel` and `ReadCommitment` functions.

Testing showed that with large but valid asset amounts, the channel
custom records buffer size exceeded `tlv.MaxRecordSize`. To address
this, we have two options: increase the max record size for channel data
or dynamically set the max read size. Given that the data is already in
memory, the latter approach is more robust.