multi: separate test/mock code from JSON models
To avoid having shared non-test code that uses testing dependencies
(such as "require") that aren't compatible with Tinygo, we move all mock
code out of production packages.
This would cause circular package dependencies if we hadn't moved all
affected test code to their own packages as well in previous commits.
To be able to re-use the JSON models that we use for the test vectors in
tinygo code, we also move those to their own package.