sqlc/migrations: add new unique key to assets table
Fixes #691 by adding a unique constraint that will actually work. The
assumption of the constraint is that there will never be an asset of the
same asset ID with the same script key in the same anchor transaction
output. This would be impossible anyway because the asset leaves would
collide within the asset-level MS-SMT tree.
So this unique key is safe to use as an upsert detection mechanism.