tapchannel: factor in breach csv when in deriveCommitKeys
When we're handling a breach, we have two CSVs to worry about: our csv,
and the csv of the remote party.

Before this commit, we would always use our csv. However, this is wrong
when we're making a script key for the to local output of the remote
party. Instead, we need to be aware of the breach csv, which is the csv
of their local output.

We use a newly added optional field to select between the correct
values.