tapdb: pass migration versions into MigrationTarget closure
This commit modifies the database migration procedure. Before performing
any actual migration, we first compute the current database migration
version and the maximum known migration version (the highest version we
could migrate to). This information is logged. We then pass this
information into the `MigrationTarget` closure, allowing it to decide
whether and how to migrate. This enables us to conditionally create a
database backup if a migration upgrade is possible.