tapdb: implement basic migration downgrade protection
In this commit, we implement basic migration downgrade protection. This
ensures that tapd will refuse to downgrade versions if a migration
happened in the latest software version.

We achieved this by hard coding the latest version as a constant. Then
when we go to migrate, we check this version against the latest version
stored in the database. If the hard coded version is less than the
latest version, then we conclude this is a downgrade attempt.