tapgarden: fetch finalized batch via file archiver
In this commit, we update listBatches to fetch assets from finalized
batches via the proof file archiver instead of the DB. Given the batch
seedlings, we fetch issuance proofs and decode the asset in its genesis
state. We also verify that we can recompute the genesis output script
with the fetched assets.