proof+tapdb: add FetchIssuanceProof to Archiver
In this commit, we add a new method to proof.Archiver to support
querying for proofs with only the asset ID and anchor outpoint. This
is useful when fetching issuance proofs in order to display genesis
assets, as assets in the DB can be mutated if they are reanchored as
passive assets (participate in a transfer as change). This is only
implemented for the FileArchiver, as we can use proof file paths to
efficiently fetch only the issuance proofs.