multi: add decimal display as new TLV field
With this commit we add the decimal display as a new TLV field on the
meta reveal. We'll still encode it within the JSON meta data if the meta
data is of type JSON. For all other types we can now also carry the
value, to make sure we don't put any type constraints on future asset
mints (in case issuers want to use a different format than JSON).

Since we're using TLV, old meta reveals should still be read correctly
and the default value of 0 is assumed for any meta reveal that doesn't
have the explicit record set.
We'll add a unit test for that in the next commit.