tapchannel: fail early if we'd send 0 assets
Fixes #1006.

In this commit we make sure that we'll never end up attempting to send a
zero asset amount HTLC, which would then fail further down the line.
A zero amount can happen if the resolution of asset unit to satoshis is
insufficient and one unit costs more than the satoshi amount that should
be represented in the HTLC.
The main mitigation for this is on the issuer side that they make sure
the decimal display value is high enough and one asset unit represents a
fraction of a cent (and is therefore close to being equal to one
satoshi).