server+tapchannel: make aux leaf creator stateless
This fixes a startup dependency issue between lnd and tapd where both
wait for each other to become ready.
Because the aux leaf creator doesn't depend on any other components that
need to be started first, and also doesn't carry any state by itself, we
can declare it to be fully stateless and not block on server startup
before being able to use it.