tapchannel: fix split root when initiator has 0 balance
This fixes a bug that lead to a panic in the following edge case:
 - The initiator doesn't have any asset balance (everything was moved to
   the responder).
 - An HTLC is created that sends some assets back to the initiator.

The problem that created is that because the initiator doesn't have any
non-HTLC balance, no allocation was created with SplitRoot=true, which
lead to a nil pointer panic in the logic that extracts the split root
locator when creating a split commitment.

This commit fixes the problem by assigning the first HTLC that pays to
the initiator with SplitRoot=true, if the initiator has zero balance.