tapchannel: ensure genesis proofs for input assets already known
With this commit, we partially address #1026 by ensuring that when we
receive an incoming funding request, then we've already verified the
genesis proofs of the incoming asset.

If we haven't verified the proofs yet, then we'll try to fetch from a
remote universe. Otherwise, this check will be a noop.