tapchannel: fix split root when initiator has 0 balance
A previous bug fix was wrong... In the case where an HTLC sends all the
balance from the initiator to the recipient, then the initiator has zero
balance _and_ there is no HTLC going to the initiator. So the root split
still becomes nil and we panic.
So we re-declare the rule for the case where the initiator has zero
balance: If the initiator doesn't have any balance to carry the split
root, then the very first HTLC will carry the split root. If there is no
HTLC, it means all the balance is on the recipient's side and there is
no split (hence no split root required).