cmd+rpcserver+taprpc: filter unconfirmed assets by default
By default, we only want to show assets in the asset list that are ready
to be spent. So we don't want to include not yet confirmed freshly
minted assets (those are the only assets that are materialized in the
asset table while not being confirmed yet).
We add a new RPC and CLI flag that allows the user to show the assets on
explicit request.