tapdb: add proof_delivery_complete and position to transfer outputs
Add the `proof_delivery_complete` and `position` columns to the
`asset_transfer_outputs` table. The `position` column indicates the
position of the output in the transfer output list.

The position and anchor outpoint uniquely identify a transfer output.

This change updates the row insertion and retrieval SQL statements for
this table. Additionally, new SQL statements are included for modifying
these new columns.