tapchannel: use defer to ensure lease unlocked if funding failure
Once we call `fundVirtualPacket`, we've locked inputs assuming it
succeeds. If for w/e reason, the remote peer fails to send us the first
ack (before the call to `completeChannelFunding`), then before this
commit, we would fail to unlock the inputs. Only after the call to
`completeChannelFunding` would we unlock the UTXOs.

In this commit, we use two new defers to ensure that we always unlock
the UTXOs if the pre-setup fails, and also if `completeChannelFunding`
fails.