itest: add basic minimal functionality tests for tchrpc
With this commit we make sure that the taproot asset channel
functionality RPCs are correctly configured and return the expected
error when being called.
We can't actually test any of their intended functionality because that
would require us to run within a litd context.
But the integration tests in the litd repo make sure these RPCs work as
expected.