tapdb: generalise QueryAssetTransfers arg flag unconf_only
This commit renames and generalizes the `unconf_only` argument in the
`QueryAssetTransfers` SQL query to `pending_transfers_only`. The flag
now returns pending transfers with unconfirmed anchoring transactions or
undelivered transfer output proofs.

Additionally, this update includes the anchor transaction block hash in
the query's returned values. This allows users to determine if a
transfer's anchoring transaction has been confirmed on-chain, providing
context for why the transfer is pending.

As a result of this change, ExportLog.PendingParcels will now also
include transfers that are pending due to undelivered proofs, even if
their associated anchoring transactions have already been confirmed
on-chain.