tapfreighter+tapdb: rename ConfirmParcelDelivery to LogAnchorTxConfirm
This commit renames the ExportLog method from ConfirmParcelDelivery to
LogAnchorTxConfirm. The purpose of this change is to clarify that the
method does not confirm parcel delivery to disk, as transfer proofs may
still be pending. Instead, it updates the send package on disk to
indicate that the transfer anchoring transaction has been confirmed
on-chain.