migrations: re-apply migration 14 as 23.
Due to a git branch cherry-pick mishap, the migration number 14 wasn't
included in a released version and therefore wasn't applied on some
systems. Because the tables in that migration file were only being used
with a recently released version (v0.4.0/v0.4.1), this wasn't noticed
for a long time.
We fix this by re-applying the same migration as number 23, slightly
modified so that it's a no-op if the tables and entries already exist on
a system.