scripts+tapdb: do sqlc required replacement in script
Our SQL scripts on disk were neither SQLite nor Postgres compatible, we
used a version suited for sqlc to generate the correct data types we
wanted, then did some type replacements in-memory when actually applying
the migration files.
This is somewhat dangerous as it will lead to situations where if a user
applies a migration file manually, they'll end up with table primary
keys that ARE NOT auto incrementing.

To fix this issue, we make sure that we always write fully SQLite
compatible SQL code and only apply fixes for the sqlc code generation
directly before generating the code.