universe: insert group anchors immediately
This commit fixes an issue with large sets of asset issuances:
When there are more assets in a group than the current batch sync limit
(200), then there's a chance that the group anchor (the one asset
containing the group key reveal record) isn't in the first batch. So we
wouldn't be able to insert the first batch at all, since those assets
would all reference a group key that isn't yet known.

To avoid this problem, we insert any proofs with group key reveal
records immediately and only start batching the others after going
through all of them.
This means we'll keep more assets in memory and can start to process
them later. But at least the process won't fail anymore.