rfq+rfqmsg: parse accept type based on outgoing request
This is a preparatory commit. Because the accept wire message will look
the same for both buy and sell accepts, we will no longer be able to
distinguish between them based on the content alone.
But that isn't necessary anyway, because we can just map it to the
request type, which is identified by the unique ID in both request and
response.
And because we just made sure each response is mapped to its request, we
also remove two old TODOs that mentioned exactly that.