github: use litd staging branch for itests
We actually want to run the litd itests on the staging branch that has
the custom channel tests.
We also can speed up the tests by removing the yarn build for the static
files, since we don't need those for the tests (an empty index.html just
needs to exist for the build to succeed).