multi: refactor reject msg to use new TLV types
We change the reject message data encoding format to use the new TLV
libraries. This is a breaking change, because the RejectErr field will
now be encoded as a single type.
But we are going to break more things on the p2p RFQ layer within this
PR anyway, so we might as well clean that up!
We're going to set a new version (and only accept that version) in a
later commit, once we've updated all other message types.