tapchannel: improve quit handling for aux signer
In this commit, we add checks of the aux signer cancel and quit signals
at all points during aux sig batch processing when a response may be
sent. This mirrors the signal handling used in the lnwallet sigpool
worker goroutine. We also add an extra check for the job cancel signal,
since it could have been sent by LND before it could be sent from here.