rfqmsg: generate ID for SCID alias into custom range
A recent change to the lnd SCID alias RPCs requires us to generate SCID
aliases in a certain range. Because we derive the alias from the
randomly generated RFQ ID, we need to make sure we derive a random ID
that can be successfully transformed into a valid SCID alias.