tapchannel: adjust timing for quit and cancel test
Testing aux signer quit behavior involved building a job queue and then
checking that some portion of the jobs were skipped, and that the last
job was not processed. This led to racy behavior if the job queue was
too short, or the delay after sending a signal was too long. Using a
larger queue and shorter delay reduces the odds of a race here.