tapd: fix L1 assets balance calculation
This commits modifies the SQL queries `QueryAssetBalancesByGroup` and
`QueryAssetBalancesByAsset` to allow an `exclude_key` parameter to
exclude assets with that key from the balance calculation.

`assetsStore.QueryBalancesByAsset` and
`assetsStore.QueryAssetBalancesByGroup` use this parameter to exclude
assets that are specifically used for funding custom channels. The
balance of these assets should be reported through the L2 channel
balance and not through the L1 asset balance.