tapchannelmsg: expand fields of ContractResolution for 2nd level sweeps
In this commit, we expand the fields of the `ContractResolution` struct
to account for second level sweeps. When we have an HTLC that needs to
go to the second level, we'll need to create 2 vPkts: one for a direct
spend from the commitment txns with the second level txn, and the other
for the spend from the sweeping transaction.

We can construct that transaction early, but can only sign for it once
we know the sweeping transaction. Therefore, we need to keep some extra
data for a given resolution, namely the tapTweak and the control block.