tapchannel: add 2nd level information to tapscriptSweepDesc
In this commit, we add two fields to the tapscriptSweepDesc struct
related to 2nd level HTLCs. First, we add the second level sig index,
which is needed to be able to know where to insert our signature after
we generate it. We also add an optional lnwallet.AuxSigDesc, which
contains the remote party's signature, and also the sighash needed for
it and a sign desc to generate it with.