tapchannel: add sweep desc gen for local HTLC sweeps
In this commit, we add sweep desc gen for local HTLC sweeps. In this case, we 'll need to sweep from both the first and second level. As a result, we'll need some additional information such as where to put the remote party's 2nd level sig. For the success sweeps, we'll also need a pre-image place holder, as we may only learn that once things are confirmed and the pre-image has been revealed off or on chain.