tapchannel: add blobWithWitnessInfo struct
In this commit, we add `blobWithWitnessInfo` which couples the raw
resolution TLV blob with some additional information such as the input
that carried the blob, and also preimage information. Preimage
information is needed to know where to insert the preimage in the
witness once we learn about it on chain.