tapchannel: update sweepContracts for HTLC sweeps
In this commit, we add some intermediate functions and types needed to
properly handle HTLC sweeps.

We now use the preimage info added in a prior commit to make sure that
once we learn of the preimage (after vPkt creation, but before publish),
we'll properly insert it in the correct location.

In sweepContracts, when we go to create the new anchor change output,
we'll make sure to omit any second level transactions, as they already
have a destination location specified.

Direct spends are spends directly from the commitment transaction. If we
don't have any direct spends, then we don't need to make a change addr,
as second level spends are already bound to an anchor output.