tapchannel: update registerAndBroadcastSweep for HTLC sweeps
In this commit, we update `registerAndBroadcastSweep` to be able to
handle HTLC sweeps.

First, we only need to set the anchor for first level sweeps.

We weren't able to sign the 2nd level sweeps earlier as we didn't know
the txid of the transaction that swept them. Now that we're about to
broadcast, we know the sweeping transaction so we can set the prevID
properly, then sign the sweeping transaction.

If the sweeper is broadcasting _just_ a second level txn with an asset,
then we won't have an extra change output. This transaction also already
has its internal key bound.