GitHub: switch litd testing branch
With this commit we switch the testing branch to the currently unmerged
'update-to-lnd-18-4' branch which will soon be in master.
But because that branch depends on the commits in this PR, we first have
to point to it, get this PR merged, then merge update-to-lnd-18-4, then
change this to 'master'.