asset: introduce NewSpecifier constructor for asset.Specifier
This commit adds a new constructor, `NewSpecifier`, for creating an
`asset.Specifier`. The constructor accepts an asset ID, a group public
key, or a group key to identify the asset. It includes a
`mustBeSpecified` parameter, which, when set to true, enforces that at
least one of the asset ID, group public key, or group key must be
provided. If none are specified, an error is returned.