rfq: validate in-flight HTLC asset amounts against HTLC out msat
Update `AssetPurchasePolicy` to include an asset specifier, which is
used to select and sum asset balances in an in-flight HTLC. This ensures
the outgoing msat amount of the HTLC is not less than the incoming Tap
asset amount in msat.

This commit fixes an issue where the asset amount used was incorrectly
set on the policy instead of being derived from the in-flight HTLC.
The erroneous asset amount field in the policy has been removed.