tapdb: add universe syncer cache
This commit adds a completely new cache that is tailor made for the
queries the syncer makes (ascending, without any amounts by ID).
This additional cache has all roots (without details), so the existing
cache will only be needed for custom queries that aren't coming from the
syncer.
We should be able to down-size the existing cache quite a bit, as it
should only serve manual requests or requests coming from an UI.

We'll make the size configurable in an upcoming commit.