rfq+rpc: repurpose BuyOrder.MinAssetAmount field to AssetMaxAmt
This commit repurposes the `MinAssetAmount` field in `BuyOrder` to
`AssetMaxAmt`, defining the upper limit on the inbound asset amount
from the RFQ initiator's perspective. This change applies when the
requestor sets up an invoice to receive payment in tap asset (sending
a "buy" RFQ request).

The `AssetMaxAmt` field will be populated in a `BuyRequest` to ensure
that the counterparty can accommodate this requirement.