tapfreighter: release asset UTXOs on error
Previously we only released/unlocked the BTC level outputs we leased
from the lnd wallet when an error occurred.
But if we're still in a state where nothing has been written to disk and
the user can try again, we can safely release/unlock the asset level
UTXOs as well to avoid them being locked for 10 minutes.