tapdb: fix bug with null bool handling in db
In this commit, we fix some incorrect handling on null bool in SQL, that
looks to be mostly benign. Before this commit, instead of doing `.Bool`,
we did `.Valid`. The latter is only useful to check if something is
zero, or NULL. In this case, we can just go with the bool value
directly.