tapchannel: properly thread thru lock time for 2nd level HTLCs
In this commit, we fix an existing logic gap related to 2nd level HTLCs.
Before if we were signing a timeout for the remote party, neither of us
would properly apply the lock time to the vPkt.

In this commit, we fix it by first gaining a new `whoseCommit` param in
`FetchLeavesFromCommit`. We then use that to decide when to use a non
zero CLTV timeout.