tapchannel: ensure the root commit asset has a valid witness
In this commit, we fix a very subtle issue related to split witnesses
and root assets.

Before this commit, when we went to sign a second level txn, which is a
split, we would create a virtual txn, which will commit to the _entire_
witness of the root asset in the vIn. Note that this bypasses the normal
segwit encoding logic, as that currently doesn't apply to the root asset
encoded in the split commit proof.

In the future, we may want to expand the segwit encoding semantics to
root assets after careful consideration.

In this commit, we fix a bug that would cause an invalid 2nd level sig
by ensuring that before we create the output commitments, we update the
root asset witness in each split, to the funding witness script, which
is just OP_TRUE.