tapchannel: add an extra budget for each input w/ a blob
In this commit, we fix an issue that would cause lnd to not broadcast
sweeps of HTLCs due to their small value. Before we didn't add enough
budget to actually convince lnd to init the fee function and broadcast
the sweep.

In the future, we should do a more careful calculation here based on the
current BTC value of the asset, to make an economical decision. For now,
we just increase the amt based on the amt of inputs we have. This gives
lnd a fee budget to use for bumping.