tapfreighter: de-duplicate passive asset creation
Because we now check all active and passive packets for uniqueness in
their commitment keys, this issue has surfaced. If we're spending
multiple different assets from the same on-chain input commitment, we
used to create duplicate passive assets.
That wasn't a problem until now because within the trees everything
would just be overwritten and collapsed back to a single asset. But
because have the uniqueness check on the virtual packet slice level,
this duplication became apparent.