rfq: fix precision issue in HTLC compliance check
Due to the integer nature of asset units, when converting to
milli-satoshi and back, we might be off by a fraction of an asset unit
when comparing to the milli-satoshi amount of an HTLC.
We'll add a single asset unit to the reported inbound amount when
checking in-out amount compliance to allow for those asset rounding
imprecisions. In a real-world scenario where an asset has a decimal
display value of 6, this would result in an imprecision of fractions of
cents and be earned back by the spread on the actual conversion itself.