rfqmsg: use transfer type for buy/sell request classification
Update the buy/sell classification logic for incoming requests to rely
on the new transfer type field.

When the requesting peer attempts to pay an invoice using a Tap asset,
they are "selling" the Tap asset to the edge node. Conversely, when the
requesting peer attempts to receive a Tap asset as payment to settle an
invoice, they are "buying" the Tap asset from the edge node.