tapchannel: tweak HTLC script key internal key to enforce uniqueness
This commit tweaks the internal key of the asset-level script key with
the HTLC index to enforce uniqueness of script keys across multiple
HTLCs with the same payment hash and timeout (MPP shards of the same
payment).
The internal key we tweak is the revocation key. So we must also apply
that same tweak when doing a revocation/breach sweep transaction. But
since that functionality is not yet implemented, we'll just leave a TODO
in a follow-up commit for that.
Therefore, no tweak will currently need to be applied at _sign_ time.