rpc+tapgarden: FundBatch and ListBatches return unsealed PSBTs
Add a new PSBT field to the `UnsealedAssets` RPC message type, which
contains the byte-serialized PSBT equivalent of the group virtual
transaction for unsealed assets. As a result, the `FundBatch` and
`ListBatches` RPC endpoints now return group virtual PSBTs.

Include logic to generate the group virtual PSBT. The PSBT is unsigned
and is provided to allow signing with an external cold private key.