multi: add trace log outputs
This commit adds a couple of trace outputs that were helpful in finding
two bugs fixed in this series of commits.

We also increase the number of levels the spew logger is allowed to
print, so we can see a bit more. We need this limit to avoid endless
loops in self-referencing structs such as proofs.