CI: update release notes template re timestamps
The release notes template has been updated to improve the paragraph on
timestamps. It now correctly specifies both ots artifacts and how to use
them to verify the timestamps of the manifest and its signature.

Secondly, the template now explains how the timestamp of the _git tag_
can be verified by using the OpenTimestamps git wrapper.