tapfreighter: trim altLeaves from passive vPkts
In this commit, we update the passive asset handling to not consider
altLeaves as assets. Specifically, any altLeaf in an input Tap
commitment should be set in the vInput for a passive asset vPkt, but
should not have its own vPkt where it is an input. Overall, input
altLeaves do not need to be re-anchored.