tapgarden: add new function newVerboseBatch
Introduced a `newVerboseBatch` function to create a verbose mint
batch from a regular mint batch. This functionality is extracted
from the existing list batch logic as part of a refactor.

The function will be used in a future commit to include a verbose
batch in the response of the FundBatch RPC endpoint.