multi: add new v2 version of GKR based on Pedersen commitments
In this commit, we add a new version of the new more flexible GKR type.
This version uses an `OP_CHECKSIG` with an un-spendable public key. We
derive such a public key by using the Pedersen commitment of the asset
ID, or blank message (for the dummy leaf case). As a Pedersen commitment
uses an aux generator H, that no one knows the private key to, it's
impossible to generate a signature using the commitment point.

An informal sketch of a proof by contradiction goes something like:
  * Starting with a Pedersen commitment, P = m*G + r*H.
  * Let's assume that we can actually sign with P. This means that
    there's some x, where P = x*G that we know.
  * If we re-write x in terms of our commitment, we have: P = m*G + r*H.
  * Wlog, we'll simplify by removing `r`, the blinding factor, so: P =
    m*G + H.
  * In terms of private values we have x*G = m*G + H
  * If we solve for H, then we have H = (x-m)*G
  * If this is a Pedersen commitment, then the private key of H must be
    unknown to all parties.
  * If we can sign with P, then this means that we know the value (x-m).
  * However, since this is a Pedersen commitment, it's impossible to
    know the value `x`, which is a function of the private value H of
    the pedersen commitment (h*H = (x-m)*G. This is a contradiction, as
    we see the that if we know x-m, then we also know h.