tapchannel: add asset htlc validation function
In this commit we utilize the previously defined helper in order to
define a new htlc validation function. The validation currently includes
checking that the asset ID of the rfq and the HTLCs always match. This
is done as an extra strict check to guard against HTLC and RFQ asset
specifier mismatch, which can lead to malicious behavior where a quote
for a different asset is being accounted for when accepting an asset
HTLC.