tapdb: always store result of querySyncStats
Our universe stats could run into a case where if in an un-cached state
the db query takes longer than the default client timeout (usually 30s)
then we'd fail the query and return early. This would cause the next
client to also receive an error instead of the cached result, making
this codepath stuck in an endless failing query. With this change we
parallelize the db query and we always wait for the result in order to
store it in cache. We may fail the function earlier, but the result will
always be retrieved & persisted.