tapdb: add TestUniverseStatsAsyncCache test
With this commit we add a simple test for the asynchronous cache
population that was introduced previously. We make a single RPC call to
the aggregate stats endpoint with a client timeout that is too small.
This leads to an RPC failure, so we check whether the cache was updated
in the background.