tapchannel: update AuxInvoiceManager tests for new asset ID check
For test cases where an RFQ does not exist we fail earlier, so we need
to update the expected behavior on the tests. We also need to be more
strict when creating dummy RFQs, as now the request with the asset
specifier needs to be present, for the new checks to take place.